Cambios_nuevos
Se agregan los estados, para el uso de la función "ACTIVAR", "DESACTIVAR", uso de id con string para los dto, manejo de aplicación por capas.
