+++ ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/AccesoADatos/model/NivelEntity.java
@@ -1,0 +1,28 @@
+package org.unicauca.modulorubricacriterio.AccesoADatos.model;
+
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "Nivel")
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class NivelEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long idNivel;
+
+    @ManyToOne
+    @JoinColumn(name = "idCriterio")
+    @JsonBackReference
+    private CriterioEntity criterio;
+
+    private String nivelDescripcion;
+    private String rangoNota;
+
+    
+}
