+++ ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/AccesoADatos/repository/NivelRepository.java
@@ -1,0 +1,7 @@
+package org.unicauca.modulorubricacriterio.AccesoADatos.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.unicauca.modulorubricacriterio.AccesoADatos.model.NivelEntity;
+
+public interface NivelRepository extends JpaRepository<NivelEntity,Long> {
+}
