+++ ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/Fachada/estados/IEstados.java
@@ -1,0 +1,8 @@
+package org.unicauca.modulorubricacriterio.Fachada.estados;
+
+import org.unicauca.modulorubricacriterio.Fachada.validacionEstados.EstadosController;
+
+public interface IEstados {
+    Resultado activar(EstadosController estadosController);
+    Resultado desactivar(EstadosController estadosController);
+}
