+++ ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/Fachada/estados/Resultado.java
@@ -1,0 +1,10 @@
+package org.unicauca.modulorubricacriterio.Fachada.estados;
+
+/**
+ * @boolean cambioPermitido: Indica si el cambio de estado es permitido o no.
+ * @String mensaje: Mensaje que describe el resultado del cambio de estado.
+ */
+
+public record Resultado(boolean cambioPermitido, String mensaje) {
+
+}
