+++ ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/Fachada/validacionEstados/EstadosEnum.java
@@ -1,0 +1,5 @@
+package org.unicauca.modulorubricacriterio.Fachada.validacionEstados;
+
+public enum EstadosEnum {
+    ACTIVO, INACTIVO
+}
