--- ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/dto/NivelDTO.java
@@ -1,15 +1,0 @@
-package org.unicauca.modulorubricacriterio.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class NivelDTO {
-    private Long idNivel;
-    private Long id_criterio;
-    private String nivelDescripcion;
-    private String rangoNota;
-}
