--- ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/exception/excepcionStructure/Error.java
@@ -1,41 +1,0 @@
-package org.unicauca.modulorubricacriterio.exception.excepcionStructure;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-@Data
-@Builder(toBuilder = true)
-@NoArgsConstructor
-@AllArgsConstructor
-
-public class Error {
-    /**
-     * Código de error de la aplicación
-     */
-    private String codigoError;
-
-    /**
-     * Mensaje de error generado
-     */
-    private String mensaje;
-
-    /**
-     * Código de estatus http
-     */
-    private Integer codigoHttp;
-
-    /**
-     * Url de la petición que generó el error
-     */
-    @Accessors(chain = true)
-    private String url;
-
-    /**
-     * Método de la petición que generó el error
-     */
-    @Accessors(chain = true)
-    private String metodo;
-}
