--- ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/exception/exceptionOwn/GestionModuloRuntimeException.java
@@ -1,16 +1,0 @@
-package org.unicauca.modulorubricacriterio.exception.exceptionOwn;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.unicauca.modulorubricacriterio.exception.excepcionStructure.CodigoError;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public abstract class GestionModuloRuntimeException extends RuntimeException {
-
-  protected CodigoError codigoError;
-  public abstract String formatException();
-
-}
