--- ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/repository/CriterioRepository.java
@@ -1,9 +1,0 @@
-package org.unicauca.modulorubricacriterio.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import org.unicauca.modulorubricacriterio.model.CriterioEntity;
-
-@Repository
-public interface CriterioRepository extends JpaRepository<CriterioEntity, Long> {
-}
