--- ModuloRubricaCriterio/src/main/java/org/unicauca/modulorubricacriterio/repository/NivelRepository.java
@@ -1,7 +1,0 @@
-package org.unicauca.modulorubricacriterio.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.unicauca.modulorubricacriterio.model.NivelEntity;
-
-public interface NivelRepository extends JpaRepository<NivelEntity,Long> {
-}
