+++ ModuloRubricaCriterio/src/main/resources/application.properties
@@ -30,6 +30,6 @@ spring.jpa.show-sql=true
 
 #el siguiente comando inicia la base de datos cuando inicia la aplicacion
 #si ya esta creada no hace nada
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=update
